% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{CI_metrics}
\alias{CI_metrics}
\title{CI_metrics}
\usage{
CI_metrics(TP, TN, FP, FN, a = 0.05)
}
\arguments{
\item{TP}{- scalar}

\item{TN}{- scalar}

\item{FP}{- scalar}

\item{FN}{- scalar}

\item{a}{- alpha level}
}
\value{
list
}
\description{
CI_metrics
}
