% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cohens_kappa}
\alias{cohens_kappa}
\title{cohens_kappa}
\usage{
cohens_kappa(TP, TN, FP, FN)
}
\arguments{
\item{TP}{- [scalar] True Positives}

\item{TN}{- [scalar] True Negatives}

\item{FP}{- [scalar] False Positives}

\item{FN}{- [scalar] False Negatives}
}
\value{
[scalar] Cohen's Kappa statistic
}
\description{
`cohens_kappa()` is a metric scoring function that returns the Cohen's Kappa statistic. Not an exported function.
}
\details{
This function takes integer counts for True Positive `TP`, True Negative `TN`, False Positive `FP`, and False Negative `FN` values and returns the Cohen's Kappa. This statistic measure the agreement between observations and predictions. See: https://en.wikipedia.org/wiki/Cohen%27s_kappa
}
