% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_site_data.R
\name{format_site_data}
\alias{format_site_data}
\title{format_site_data}
\usage{
format_site_data(dat, N_sites, train_test_split, background_site_balance)
}
\arguments{
\item{dat}{- data.frame}

\item{N_sites}{- scaler}

\item{train_test_split}{- scaler}

\item{background_site_balance}{- scaler}
}
\value{
- list of various ways to arrange site data and mean/sd of data
}
\description{
format_site_data
}
