% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_data.R
\name{get_sim_data}
\alias{get_sim_data}
\title{get_sim_data}
\usage{
get_sim_data(sites_var1_mean = 50, sites_var1_sd = 10,
  sites_var2_mean = 3, sites_var2_sd = 2, backg_var1_mean = 100,
  backg_var1_sd = 20, backg_var2_mean = 6, backg_var2_sd = 3,
  site_samples = 400, N_site_bags = 0, background_site_balance = 1,
  test_train_split = 0.75)
}
\arguments{
\item{sites_var1_mean}{- scaler}

\item{sites_var1_sd}{- scaler}

\item{sites_var2_mean}{- scaler}

\item{sites_var2_sd}{- scaler}

\item{backg_var1_mean}{- scaler}

\item{backg_var1_sd}{- scaler}

\item{backg_var2_mean}{- scaler}

\item{backg_var2_sd}{- scaler}

\item{site_samples}{- scaler}

\item{N_site_bags}{- scaler}

\item{background_site_balance}{- scaler}

\item{test_train_split}{- scaler}
}
\value{
- data.frame of simulated site data
}
\description{
get_sim_data
}
