% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{K_corrplot}
\alias{K_corrplot}
\title{K_corrplot}
\usage{
K_corrplot(K, train_data, clusters = 4)
}
\arguments{
\item{K}{- Matrix}

\item{train_data}{- data.frame}

\item{clusters}{- integer}
}
\value{
- a plot
}
\description{
K_corrplot
}
