% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMERR_funs.R
\name{KRR_logit}
\alias{KRR_logit}
\title{KRR_Logit}
\usage{
KRR_logit(K, y, lambda)
}
\arguments{
\item{K}{- matrix}

\item{y}{- vecotr}

\item{lambda}{- scaler}
}
\value{
list: `pred` - predicted probabiity of positive class, `alphas` - estimated coefficients
}
\description{
KRR_Logit
}
