% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMERR_funs.R
\name{KRR_logit_optim}
\alias{KRR_logit_optim}
\title{KRR_logit_optim}
\usage{
KRR_logit_optim(K, presence, lambda, maxiter = 100, tol = 0.01)
}
\arguments{
\item{K}{- matrix}

\item{presence}{- vector}

\item{lambda}{- scaler}

\item{maxiter}{- integer}

\item{tol}{- double}
}
\value{
list: `pred` - predicted probabiity of positive class, `alphas` - estimated coefficients
}
\description{
KRR_logit_optim
}
