% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMERR_funs.R
\name{KRR_logit_predict}
\alias{KRR_logit_predict}
\title{KRR_logit_predict}
\usage{
KRR_logit_predict(test_data, train_data, alphas_pred, sigma,
  dist_method = "Euclidean", progress = TRUE)
}
\arguments{
\item{test_data}{- data.frame or matrix}

\item{train_data}{- data.frame or matrix}

\item{alphas_pred}{- numeric vector}

\item{sigma}{- scaler}

\item{dist_method}{- object or character string}

\item{progress}{- logical}
}
\value{
- numeric vector - predicted probabiity of positive class
}
\description{
KRR_logit_predict
}
