% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{make_quads}
\alias{make_quads}
\title{make_quads}
\usage{
make_quads(pred, obs)
}
\arguments{
\item{pred}{- [vector] Predicted probabilities.}

\item{obs}{- [vector] Observed presence/absence as 1/0}
}
\value{
[vector] TP, FP, TN, FN names vector
}
\description{
make_quads
}
