% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{metrics}
\usage{
metrics(TP, TN, FP, FN)
}
\arguments{
\item{TP}{- scalar}

\item{TN}{- scalar}

\item{FP}{- scalar}

\item{FN}{- scalar}

\item{a}{- alpha level}
}
\value{
list - metrics
}
\description{
metrics
}
