% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_predict_funs.R
\name{rescale_sim_raster}
\alias{rescale_sim_raster}
\title{rescale_sim_raster}
\usage{
rescale_sim_raster(rast, mean, sd)
}
\arguments{
\item{rast}{}

\item{mean}{}

\item{sd}{}
}
\value{
[raster]
}
\description{
rescale_sim_raster
}
