% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_predict_funs.R
\name{scale_prediction_rasters}
\alias{scale_prediction_rasters}
\title{scale_prediction_rasters}
\usage{
scale_prediction_rasters(pred_var_stack, params, verbose = 1)
}
\arguments{
\item{pred_var_stack}{}

\item{params}{}

\item{verbose}{}
}
\value{
[raster]
}
\description{
scale_prediction_rasters
}
