% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_predict_funs.R
\name{split_raster_stack}
\alias{split_raster_stack}
\title{split_raster_stack}
\usage{
split_raster_stack(rast_stack, ppside, ngb, split = TRUE)
}
\arguments{
\item{rast_stack}{}

\item{ppside}{}

\item{split}{}
}
\value{
[list] a list of raster chips
}
\description{
Internal function to used by `KLR_raster_predict()` split input rasters into overlapping blocks for prediction.
}
\details{
`split_raster_stack()` takes a raster stack, creates a grid of `ppside` pixels, adds a collar, clips out each raster, and returns a list
}
