% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLR_funs.R
\name{tri_swap}
\alias{tri_swap}
\title{tri_swap}
\usage{
tri_swap(m)
}
\arguments{
\item{m}{- Matrix}
}
\value{
- Matrix
}
\description{
`tri_swap()` is an internal helper function used to flip the sides of a matrix
}
