% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMERR_funs.R
\name{build_K}
\alias{build_K}
\title{build_k}
\usage{
build_K(y1, y2 = y1, sigma, progress = TRUE, dist_method)
}
\arguments{
\item{y1}{- Matrix}

\item{y2}{- Matrix}

\item{sigma}{- scaler}

\item{progress}{- logical}

\item{dist_method}{- option passed to get_k for `dist_method`}
}
\value{
- matrix K
}
\description{
build_k
}
