% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMERR_funs.R
\name{build_K}
\alias{build_K}
\title{build_k}
\usage{
build_K(y1, y2 = y1, sigma, progress = TRUE, dist_method)
}
\arguments{
\item{y1}{- [NxP] Matrix of data from bag i}

\item{y2}{- [NxP] Matrix of data from bag j}

\item{sigma}{- [scaler] smoothing hyperparameters for RBF kernel}

\item{progress}{- [logical] False = no progress bar; 1 = show progress bar}

\item{dist_method}{- [object] distance method from proxy package, e.g. proxy::pr_DB$get_entry("Euclidean")}
}
\value{
- matrix K
}
\description{
build_k
}
