% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{CM_quads}
\alias{CM_quads}
\title{CM_quads}
\usage{
CM_quads(dat, threshold = 0.5)
}
\arguments{
\item{dat}{- [data.frame] Table with two columns, "pred" and "presence". "pred" is the predicted probability. "presence" is the observed presence/absence as 1/0.}

\item{threshold}{- [scalar or vector] a scalar or vector of one or more thresholds at which to evaluate the Confusion Matrix quadrants.}
}
\value{
[data.frame] Confusion Matrix quatrants at one or more threshold values.
}
\description{
CM_quads
}
