% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cohens_kappa}
\alias{cohens_kappa}
\title{cohens_kappa}
\usage{
cohens_kappa(TP, TN, FP, FN)
}
\arguments{
\item{TP}{- [scalar] True Positives}

\item{TN}{- [scalar] True Negatives}

\item{FP}{- [scalar] False Positives}

\item{FN}{- [scalar] False Negatives}
}
\value{
[scalar] Cohen's Kappa
}
\description{
cohens_kappa
}
