% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{cohens_kappa}
\alias{cohens_kappa}
\title{cohens_kappa}
\usage{
cohens_kappa(TP, TN, FP, FN)
}
\arguments{
\item{TP}{- scalar}

\item{TN}{- scalar}

\item{FP}{- scalar}

\item{FN}{- scalar}
}
\value{
scalar k
}
\description{
cohens_kappa
}
