% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_predict_funs.R
\name{crop_raster_collar}
\alias{crop_raster_collar}
\title{crop_raster_collar}
\usage{
crop_raster_collar(pred_rast, ngb, cols, rows)
}
\arguments{
\item{pred_rast}{[raster] predicted raster}

\item{ngb}{[scalar] square neighborhodd size in cells}
}
\value{
[raster]
}
\description{
crop_raster_collar
}
