% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_site_data.R
\name{format_site_data}
\alias{format_site_data}
\title{format_site_data}
\usage{
format_site_data(dat, N_sites, train_test_split, background_site_balance,
  sample_fraction)
}
\arguments{
\item{dat}{- [data.frame] A data.frame of presence and absence records. Column "presence" must contain presence/absence as 1/0, and column "SITENO" contains the grouping variable.}

\item{N_sites}{- [scaler] The number of sites to randomly select}

\item{train_test_split}{- [scaler] a float from 0 to 1 indicating the percent of N_sites to be used as training data.}

\item{background_site_balance}{- [scaler] Integer > 0 indicating how many background groups per site group}
}
\value{
- list of various ways to arrange site data and mean/sd of data
}
\description{
format_site_data
}
