% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLR_funs.R
\name{get_k}
\alias{get_k}
\title{get_k}
\usage{
get_k(y1, y2, sigma, dist_metric = "euclidean")
}
\arguments{
\item{y1}{- [NxP] Matrix of data from bag i}

\item{y2}{- [NxP] Matrix of data from bag j}

\item{sigma}{- [scaler] smoothing hyperparameters for RBF kernel}

\item{dist_metric}{[character] One of the distance methods from rdist::cdist. Default = "euclidean". see ?rdist::cdist}
}
\value{
Matrix G
}
\description{
`get_k()` is an internal function used by `build_k()` to calcualte similarity kernel for each pair of observations
}
\details{
This function takes two data.frames or matricies, calculates the cross distance between them, and then applies a radial basis function to the resulting distance matrix.
}
