% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMERR_funs.R
\name{get_k}
\alias{get_k}
\title{get_k}
\usage{
get_k(y1, y2, sigma, dist_method = "Euclidean")
}
\arguments{
\item{y1}{- matrix}

\item{y2}{- matrix}

\item{sigma}{- scaler}

\item{dist_method}{- object or character string}
}
\value{
Matrix G
}
\description{
get_k
}
