% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMERR_funs.R
\name{get_k}
\alias{get_k}
\title{get_k}
\usage{
get_k(y1, y2, sigma, dist_method = dist_method)
}
\arguments{
\item{y1}{- [NxP] Matrix of data from bag i}

\item{y2}{- [NxP] Matrix of data from bag j}

\item{sigma}{- [scaler] smoothing hyperparameters for RBF kernel}

\item{dist_method}{- [object] distance method from proxy package, e.g. proxy::pr_DB$get_entry("Euclidean")}
}
\value{
Matrix G
}
\description{
get_k
}
