% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLR_funs.R
\name{KLR}
\alias{KLR}
\title{KLR}
\usage{
KLR(K, presence, lambda, maxiter = 100, tol = 0.01, verbose = 1)
}
\arguments{
\item{K}{- [NxN] Mean embedding kernel matrix}

\item{presence}{- [vector] or presence = 1 or absence = 0}

\item{lambda}{- [scaler] Ridge regularization parameter}

\item{maxiter}{- [integer] Maximum iterations for IRLS algorithm}

\item{tol}{- [double] The convergence tolerance}

\item{verbose}{- [scaler] 0 = No notice; 1 = Reports the number of steps until convergence; 2 = Reports each iteration}
}
\value{
list: `pred` - predicted probabiity of positive class, `alphas` - estimated coefficients
}
\description{
KLR
}
