% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLR_funs.R
\name{KLR_predict}
\alias{KLR_predict}
\title{KLR_predict}
\usage{
KLR_predict(test_data, train_data, alphas_pred, sigma,
  dist_method = dist_method, progress = TRUE)
}
\arguments{
\item{test_data}{- [list] Training data used to create similarity kernel matrix}

\item{train_data}{- [list] Testing data to predict class}

\item{alphas_pred}{- [vector] Numeric vector of alpha parameters from KLR function}

\item{sigma}{- [scaler] Smoothing parameter for RBF kernel}

\item{dist_method}{- [object] distance method from proxy package, e.g. proxy::pr_DB$get_entry("Euclidean")}

\item{progress}{- [logical] False = no progress bar; 1 = show progress bar}
}
\value{
- [vector] - predicted probabiity of positive class
}
\description{
KLR_predict
}
