% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_predict_funs.R
\name{KLR_predict_each}
\alias{KLR_predict_each}
\title{KLR_predict_each}
\usage{
KLR_predict_each(rast_stack, ngb, params, progress = progress)
}
\arguments{
\item{rast_stack}{}

\item{ngb}{}

\item{params}{}

\item{progress}{}
}
\value{
[raster]
}
\description{
KLR_predict_each
}
