% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_predict_funs.R
\name{KLR_raster_predict}
\alias{KLR_raster_predict}
\title{KLR_raster_predict}
\usage{
KLR_raster_predict(rast_stack, ngb, params, split = FALSE, ppside = NULL,
  progress = TRUE, parallel = FALSE)
}
\arguments{
\item{rast_stack}{}

\item{ngb}{}

\item{params}{}

\item{split}{}

\item{ppside}{}

\item{progress}{}

\item{parallel}{}
}
\value{
[raster]
}
\description{
KLR_raster_predict
}
