% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_predict_funs.R
\name{KLR_raster_predict}
\alias{KLR_raster_predict}
\title{KLR_raster_predict}
\usage{
KLR_raster_predict(rast_stack, ngb, params, cols, rows, split = FALSE,
  ppside = NULL, progress = TRUE, parallel = FALSE, output = "list",
  save_loc = NULL, overwrite = FALSE)
}
\arguments{
\item{rast_stack}{}

\item{ngb}{}

\item{params}{}

\item{split}{}

\item{ppside}{}

\item{progress}{}

\item{parallel}{}

\item{output}{[string] either 'list' or 'save'. If 'list', returns list of predicted blocks. If 'save', blocks are saved to save_loc as GeoTiff}

\item{save_loc}{[string] Location to save raster blocks (GeoTiff). Uses getwd() is NULL}

\item{overwrite}{[logical] TRUE will overwrite saved raster GeoTiffs in save_loc}
}
\value{
[raster]
}
\description{
KLR_raster_predict
}
