% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{make_quads}
\alias{make_quads}
\title{make_quads}
\usage{
make_quads(pred, obs)
}
\arguments{
\item{pred}{- [vector] Predicted probabilities.}

\item{obs}{- [vector] Observed presence/absence as 1/0}
}
\value{
[vector] TP, FP, TN, FN names vector
}
\description{
`make_quads()` Produces True Positive, False Positive, True Negative, and False Negative quantities from two binary input vectors of predicted and observed classes.
}
\details{
This function takes two binary vectors coded as 1 == presence/positive and 0 == absent/negative. The `pred` vector is the predicted binary class and `obs` vector is the true observed class. The output quantites can be used for additional metric evaluation.
}
\examples{
\dontrun{
cm <- make_quads(ifelse(test_log_pred >= 0.5, 1, 0), test_presence)
metrics(TP = cm[1], TN = cm[3], FP = cm[2], FN = cm[4])$Informedness
}

}
