% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{metrics}
\usage{
metrics(TP, TN, FP, FN)
}
\arguments{
\item{TP}{- [scalar] True Positives}

\item{TN}{- [scalar] True Negatives}

\item{FP}{- [scalar] False Positives}

\item{FN}{- [scalar] False Negatives}

\item{a}{- [scalar] alpha level}
}
\value{
[list] - list of all metrics
}
\description{
metrics
}
