% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{metrics}
\usage{
metrics(TP, TN, FP, FN)
}
\arguments{
\item{TP}{- [scalar] True Positives}

\item{TN}{- [scalar] True Negatives}

\item{FP}{- [scalar] False Positives}

\item{FN}{- [scalar] False Negatives}

\item{a}{- [scalar] alpha level}
}
\value{
[list] - list of all metrics
}
\description{
`metrics()` returns a wide range of binary class evaluation metrics based on the inputs of True Positive, True Negative, Fale Positive, and Fale Negative quantities
}
\details{
This function is a one-stop-shop to compute 50+ metric results based on the input of TP, TN, FP, and FN
}
