% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_predict_funs.R
\name{sim_trend}
\alias{sim_trend}
\title{sim_trend}
\usage{
sim_trend(cols, rows, n = 2, size = 2)
}
\arguments{
\item{cols}{}

\item{rows}{}

\item{n}{}

\item{size}{}
}
\value{
[raster]
}
\description{
sim_trend
}
