% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_predict_funs.R
\name{split_raster_stack}
\alias{split_raster_stack}
\title{split_raster_stack}
\usage{
split_raster_stack(rast_stack, ppside, ngb, split = TRUE)
}
\arguments{
\item{rast_stack}{}

\item{ppside}{}

\item{split}{}
}
\value{
[raster]
}
\description{
split_raster_stack
}
